# backend/swagger.yaml
openapi: 3.0.3
info:
  title: CyberRisk API
  description: |
    API documentation for the CyberRisk platform — an AI-powered cyber risk
    management system built with Express + Supabase.

    ## Authentication
    Most endpoints require a **Bearer token** (Supabase JWT) in the
    `Authorization` header. Obtain one via `/api/auth/login`.
  version: 1.0.0
  contact:
    name: CyberRisk Team
servers:
  - url: http://localhost:4000/api
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access_token obtained from /api/auth/login
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Something went wrong"

    # ── Auth ──────────────────────────────────
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: "securepassword123"
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user:
          type: object

    # ── Organization ──────────────────────────
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email & password
      description: Authenticates a user via Supabase and returns JWT tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      description: Creates a new user in Supabase Auth.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                description: Supabase signup response
        "400":
          description: Registration error (e.g. email already taken)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Uses a valid refresh_token to obtain a new access_token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "400":
          description: Missing refresh_token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Refresh failed / token expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Organization ────────────────────────────
...